// src/shared/enums/bill-items-category.enums.ts
export enum BillItemsCategory {
    ELECTRONICS = 'ELECTRONICS',
    CLOTHING = 'CLOTHING',
    HOME_APPLIANCES = 'HOME_APPLIANCES',
    BOOKS = 'BOOKS',
    TOYS = 'TOYS',
    BEAUTY_PRODUCTS = 'BEAUTY_PRODUCTS',
    SPORTS_EQUIPMENT = 'SPORTS_EQUIPMENT',
    FURNITURE = 'FURNITURE',
    JEWELRY = 'JEWELRY',
    AUTOMOTIVE = 'AUTOMOTIVE',
    PET_SUPPLIES = 'PET_SUPPLIES',
    OFFICE_SUPPLIES = 'OFFICE_SUPPLIES',
    GARDEN_SUPPLIES = 'GARDEN_SUPPLIES',
    MUSIC_INSTRUMENTS = 'MUSIC_INSTRUMENTS',
    VIDEO_GAMES = 'VIDEO_GAMES',
    HEALTH_PRODUCTS = 'HEALTH_PRODUCTS',
    BABY_PRODUCTS = 'BABY_PRODUCTS',
    GROCERIES = 'GROCERIES',
}